+++ CHANGELOG.md
@@ -14,6 +14,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Add `egui::plot::Plot` to plot some 2D data
 * Add `Ui::hyperlink_to(label, url)`.
 
+### Changed 🔧
+
+* Improve the positioning of tooltips.
+
+
 ## 0.9.0 - 2021-02-07 - Light Mode and much more
 
 <img src="media/0.9.0-disabled.gif" width="50%">